+++ src/components/ProposalForm/ProposalForm.jsx
@@ -23,7 +23,7 @@ import AttachFileInput from "src/components/AttachFileInput";
 import DraftSaver from "./DraftSaver";
 import { useProposalForm } from "./hooks";
 import usePolicy from "src/hooks/api/usePolicy";
-import { isAnchoring, getKeyByValue } from "src/helpers";
+import { isAnchoring, replaceBlobsByDigestsAndGetFiles, replaceImgDigestByBlob } from "src/helpers";
 import {
   PROPOSAL_TYPE_REGULAR,
   PROPOSAL_TYPE_RFP,
@@ -41,74 +41,6 @@ import ProposalGuidelinesButton from "./ProposalGuidelinesButton";
 
 /** The main goal of using a Map data structure instead of internal state here is to prevent unnecessary rerenders. We just want a way to map blobs to files objects. */
 const mapBlobToFile = new Map();
-
-const b64toBlob = (b64Data, contentType = "", sliceSize = 512) => {
-  const byteCharacters = atob(b64Data);
-  const byteArrays = [];
-
-  for (let offset = 0; offset < byteCharacters.length; offset += sliceSize) {
-    const slice = byteCharacters.slice(offset, offset + sliceSize);
-
-    const byteNumbers = new Array(slice.length);
-    for (let i = 0; i < slice.length; i++) {
-      byteNumbers[i] = slice.charCodeAt(i);
-    }
-
-    const byteArray = new Uint8Array(byteNumbers);
-    byteArrays.push(byteArray);
-  }
-
-  const blob = new Blob(byteArrays, { type: contentType });
-  return blob;
-};
-
-function replaceImgDigestByBlob(vals, mapBlobToFile) {
-  if (!vals) return { text: "", markdownFiles: [] };
-  const { description, files } = vals;
-  const imageRegexParser = /!\[[^\]]*\]\((?<digest>.*?)(?="|\))(?<optionalpart>".*")?\)/g;
-  const imgs = description.matchAll(imageRegexParser);
-  let newText = description;
-  const markdownFiles = [];
-  /**
-   * This for loop will update the newText replacing images digest by a blob and push the img object to an array of markdownFiles
-   * */
-  for (const img of imgs) {
-    const { digest } = img.groups;
-    const obj = getKeyByValue(files, digest);
-    if (obj) {
-      const urlCreator = window.URL || window.webkitURL;
-      const blobUrl = urlCreator.createObjectURL(b64toBlob(obj.payload, obj.mime));
-      mapBlobToFile.set(blobUrl, obj);
-      markdownFiles.push(obj);
-      newText = newText.replace(digest, blobUrl);
-    }
-  }
-  return { text: newText, markdownFiles };
-}
-
-/**
- * replaceBlobsByDigestsAndGetFiles uses a regex to parse images
- * @param {String} description the markdown description
- * @param {Map} map the map of blob -> file
- * @returns {object} { description, files }
- */
-function replaceBlobsByDigestsAndGetFiles(description, map) {
-  const imageRegexParser = /!\[[^\]]*\]\((?<blob>.*?)(?="|\))(?<optionalpart>".*")?\)/g;
-  const imgs = description.matchAll(imageRegexParser);
-  let newDescription = description;
-  const files = [];
-  /**
-   * This for loop will update the newDescription replacing the image blobs by their digests and push the img object to an array of files
-   * */
-  for (const img of imgs) {
-    const { blob } = img.groups;
-    if (map.has(blob)) {
-      newDescription = newDescription.replace(blob, map.get(blob).digest);
-      files.push(map.get(blob));
-    }
-  }
-  return { description: newDescription, files };
-}
 
 const ProposalForm = React.memo(function ProposalForm({
   values,
