+++ src/helpers.js
@@ -514,3 +514,80 @@ export const getDomainName = (contractorDomains, op) => {
   const domain = contractorDomains.find((domain) => domain.value === op);
   return domain ? domain.label : "";
 };
+
+/** INLINE IMAGES HELPERS */
+/**
+ * replaceBlobsByDigestsAndGetFiles uses a regex to parse images and replace blobs by files digests
+ * @param {String} description the markdown description
+ * @param {Map} map the map of blob -> file
+ * @returns {object} { description, files }
+ */
+export function replaceBlobsByDigestsAndGetFiles(description, map) {
+  const imageRegexParser = /!\[[^\]]*\]\((?<blob>.*?)(?="|\))(?<optionalpart>".*")?\)/g;
+  const imgs = description.matchAll(imageRegexParser);
+  let newDescription = description;
+  const files = [];
+  /**
+   * This for loop will update the newDescription replacing the image blobs by their digests and push the img object to an array of files
+   * */
+  for (const img of imgs) {
+    const { blob } = img.groups;
+    if (map.has(blob)) {
+      newDescription = newDescription.replace(blob, map.get(blob).digest);
+      files.push(map.get(blob));
+    }
+  }
+  return { description: newDescription, files };
+}
+
+const b64toBlob = (b64Data, contentType = "", sliceSize = 512) => {
+  const byteCharacters = atob(b64Data);
+  const byteArrays = [];
+
+  for (let offset = 0; offset < byteCharacters.length; offset += sliceSize) {
+    const slice = byteCharacters.slice(offset, offset + sliceSize);
+
+    const byteNumbers = new Array(slice.length);
+    for (let i = 0; i < slice.length; i++) {
+      byteNumbers[i] = slice.charCodeAt(i);
+    }
+
+    const byteArray = new Uint8Array(byteNumbers);
+    byteArrays.push(byteArray);
+  }
+
+  const blob = new Blob(byteArrays, { type: contentType });
+  return blob;
+};
+
+/**
+ * replaceBlobsByDigestsAndGetFiles uses a regex to parse images digests and create a new Blob
+ * @param {String} description the markdown description
+ * @param {Map} map the map of blob -> file
+ * @returns {object} { description, files }
+ */
+export function replaceImgDigestByBlob(vals, map) {
+  if (!vals) return { text: "", markdownFiles: [] };
+  const { description, files } = vals;
+  const imageRegexParser = /!\[[^\]]*\]\((?<digest>.*?)(?="|\))(?<optionalpart>".*")?\)/g;
+  const imgs = description.matchAll(imageRegexParser);
+  let newText = description;
+  const markdownFiles = [];
+  /**
+   * This for loop will update the newText replacing images digest by a blob and push the img object to an array of markdownFiles
+   * */
+  for (const img of imgs) {
+    const { digest } = img.groups;
+    const obj = getKeyByValue(files, digest);
+    if (obj) {
+      const urlCreator = window.URL || window.webkitURL;
+      const blobUrl = urlCreator.createObjectURL(
+        b64toBlob(obj.payload, obj.mime)
+      );
+      map.set(blobUrl, obj);
+      markdownFiles.push(obj);
+      newText = newText.replace(digest, blobUrl);
+    }
+  }
+  return { text: newText, markdownFiles };
+}
